working mock

+++ scripts/code_generation/jsoncan-rust/src/can_database/mod.rs
@@ -253,7 +253,7 @@ impl CanDatabase {
 pub struct DecodedSignal {
     pub name: String,
     pub value: f64,
-    pub timestamp: Option<u32>,
+    pub timestamp: Option<u64>,
     pub label: Option<String>,
     pub unit: Option<String>,
     pub signal_type: CanSignalType
